requirements.txt
expression-parser
keyring
numpy
pyocclient
pyyaml
scipy
scikit-learn
tensorflow
